Website/backend/requirements.txt
Flask==1.0.2
mysql_connector_repackaged==0.3.1
PyYAML==3.12
